% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NFLvegas.R
\name{nfl_vegas}
\alias{nfl_vegas}
\title{To extract Vegas lines}
\usage{
nfl_vegas(from = 2011, to = 2011, obs = "all")
}
\arguments{
\item{from}{The beginning year}

\item{to}{The end year}

\item{obs}{Number of observations. Default value is "all", you could put numeric values instead}
}
\description{
This function allows you to extract Vegas lines from NFL matches.
}
\examples{
nfl_vegas(from = 2011, to = 2011, obs = 5)
}
\keyword{nfl}

